address the comments
