+++ megatron/core/extensions/transformer_engine.py
@@ -1,6 +1,7 @@
 # Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 from __future__ import annotations
 
+import copy
 import dataclasses
 import enum
 import inspect
@@ -38,6 +39,7 @@ from megatron.core.tensor_parallel.layers import (
     _initialize_affine_weight_cpu,
     set_tensor_model_parallel_attributes,
 )
+from megatron.core.tensor_parallel.mappings import gather_from_tensor_model_parallel_region
 from megatron.core.tensor_parallel.random import (
     get_cuda_rng_tracker,
     get_data_parallel_rng_tracker_name,
@@ -1337,6 +1339,133 @@ class TEColumnParallelLinear(TELinear):
         """Compute weight gradients during the backward pass if delay_wgrad_compute is enabled."""
         if self.config.delay_wgrad_compute:
             super().backward_dw()
+
+
+class TELMHeadColumnParallelLinear(TEColumnParallelLinear):
+    """Wrapper for ``TEColumnParallelLinear`` used as the LM-head output projection under MXFP8.
+
+    Drop-in replacement for the ``tensor_parallel.ColumnParallelLinear`` LM head:
+    ``delay_wgrad_compute`` is forced off to mirror its no-op ``backward_dw``,
+    and ``get/set_extra_state`` match the bf16 LM head's state-dict shim. The
+    LM-head kwargs ``keep_master_weight_for_test``, ``skip_weight_param_allocation``,
+    ``defer_embedding_wgrad_compute`` buffers, and ``disable_grad_reduce`` are
+    accepted to preserve the ``ColumnParallelLinear`` signature but currently
+    raise when set non-default — TE will not support them natively, so they
+    would have to be implemented in this subclass, which has not been done yet.
+
+    Active only when ``fp8_output_proj=True`` with ``fp8_recipe='mxfp8'``.
+    """
+
+    def __init__(
+        self,
+        input_size,
+        output_size,
+        *,
+        config,
+        init_method,
+        bias=True,
+        gather_output=False,
+        stride=1,
+        keep_master_weight_for_test=False,
+        skip_bias_add=False,
+        skip_weight_param_allocation: bool = False,
+        embedding_activation_buffer=None,
+        grad_output_buffer=None,
+        is_expert: bool = False,
+        tp_comm_buffer_name: Optional[str] = None,
+        disable_grad_reduce: bool = False,
+        tp_group: Optional[torch.distributed.ProcessGroup] = None,
+    ):
+        from megatron.core.fp8_utils import is_mxfp8_output_proj_active
+
+        if not is_mxfp8_output_proj_active(config):
+            raise RuntimeError(
+                "TELMHeadColumnParallelLinear is only valid when fp8_output_proj=True, "
+                "fp8=True, and fp8_recipe='mxfp8'."
+            )
+        if keep_master_weight_for_test:
+            raise ValueError("TE output projection does not support keep_master_weight_for_test.")
+        if skip_weight_param_allocation:
+            raise ValueError("TE output projection does not support skip_weight_param_allocation.")
+        if embedding_activation_buffer is not None or grad_output_buffer is not None:
+            raise ValueError(
+                "TE MXFP8 output projection does not support defer_embedding_wgrad_compute."
+            )
+        if disable_grad_reduce:
+            raise ValueError("TE output projection does not support disable_grad_reduce.")
+
+        te_config = copy.copy(config)
+        # Match ColumnParallelLinear.backward_dw's no-op so the LM head keeps
+        # the same wgrad-timing behavior it had before this subclass existed.
+        te_config.delay_wgrad_compute = False
+
+        super().__init__(
+            input_size=input_size,
+            output_size=output_size,
+            config=te_config,
+            init_method=init_method,
+            gather_output=False,
+            bias=bias,
+            skip_bias_add=skip_bias_add,
+            is_expert=is_expert,
+            skip_weight_param_allocation=skip_weight_param_allocation,
+            tp_comm_buffer_name=tp_comm_buffer_name,
+            tp_group=tp_group,
+            stride=stride,
+        )
+
+        self.input_size = input_size
+        self.output_size = output_size
+        self.output_size_per_partition = divide(output_size, self.tp_size)
+        self.gather_output = gather_output
+        self.skip_bias_add = skip_bias_add
+        self.embedding_activation_buffer = None
+        self.grad_output_buffer = None
+        self.disable_grad_reduce = False
+        self.tp_group = self._tp_group
+
+        self._register_load_state_dict_pre_hook(
+            lambda state_dict, prefix, *args, **kwargs: state_dict.setdefault(
+                f"{prefix}_extra_state"
+            )
+        )
+
+    def get_extra_state(self):
+        # Match ColumnParallelLinear.get_extra_state (returns None) so the LM
+        # head state dict stays compatible across the bf16 / MXFP8 swap.
+        return None
+
+    def set_extra_state(self, state):
+        # Match ColumnParallelLinear.set_extra_state (ignored).
+        return
+
+    def forward(
+        self,
+        input_: torch.Tensor,
+        weight: Optional[torch.Tensor] = None,
+        runtime_gather_output: Optional[bool] = None,
+    ):
+        """Run TE MXFP8 output projection. Returns ``(output, bias)``."""
+        from megatron.core.fp8_utils import get_fp8_context
+
+        if weight is not None and weight is not self.weight:
+            raise RuntimeError("TE MXFP8 output projection does not support runtime weight.")
+
+        with get_fp8_context(self.config):
+            torch.cuda.nvtx.range_push("mxfp8_output_proj_telinear")
+            try:
+                output_parallel, output_bias = super().forward(input_)
+            finally:
+                torch.cuda.nvtx.range_pop()
+
+        gather_output = self.gather_output
+        if runtime_gather_output is not None:
+            gather_output = runtime_gather_output
+        if gather_output:
+            output = gather_from_tensor_model_parallel_region(output_parallel, group=self.tp_group)
+        else:
+            output = output_parallel
+        return output, output_bias
 
 
 class TERowParallelLinear(TELinear):
